2.8版本
2.8

+++ YLZG/YLZG/Classes/影楼业务/ERP应用/开单/离线开单/LixianOrderController.m
@@ -194,7 +194,7 @@
         self.navigationItem.rightBarButtonItem = nil;
         return;
     }
-    NSString *message = [NSString stringWithFormat:@"发送(%ld)",selectOrderArr.count];
+    NSString *message = [NSString stringWithFormat:@"发送(%ld)",(unsigned long)selectOrderArr.count];
     self.navigationItem.rightBarButtonItem = [[UIBarButtonItem alloc]initWithTitle:message style:UIBarButtonItemStylePlain target:self action:@selector(sendOrder)];
     [self.navigationItem.rightBarButtonItem setTitleTextAttributes:@{NSFontAttributeName:[UIFont systemFontOfSize:15],NSForegroundColorAttributeName:[UIColor whiteColor]} forState:UIControlStateNormal];
